Switch float to double in TimeZoneHelper
JavaScript represents all numbers as IEEE754 64-bit doubles, NOT
32-bit floats, so use doubles instead of floats to avoid any
unexpected truncation.

+++ widget/TimeZoneHelper/TimeZoneHelper.h
@@ -49,8 +49,8 @@
 - (NSArray *)allTimeZones;
 
 - (void)setTimeZoneWithName:(NSString *)name;
-- (float)timeOffsetHours;
-- (NSString *)formattedTimeForHours:(float)hours;
+- (double)timeOffsetHours;
+- (NSString *)formattedTimeForHours:(double)hours;
 - (NSString *)lookupPlaceInRegion:(NSString *)region withName:(NSString *)name;
 
 - (NSString *)myRegionCode;
